backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
pydantic==2.8.2
python-dotenv==1.0.1
openai>=1.30.0